Naming conventions for draw move tools
