+++ src/TreeContext.ts
@@ -16,8 +16,8 @@ export enum Tool {
     atomTool,
     cutTool,
     dragTool,
-    moveSingleTool,
-    moveMultiTool,
+    drawMoveSingleTool,
+    drawMoveMultiTool,
     copySingleTool,
     copyMultiTool,
     deleteSingleTool,
