+++ src/index.html
@@ -45,10 +45,10 @@
           </button>
         </div>
         <div class = "row">
-          <button type="button" onclick="setTool(moveSingleTool)" title="Move Single Tool" class="toolButton">
+          <button type="button" onclick="setTool(drawMoveSingleTool)" title="Move Single Tool" class="toolButton">
             <i class="fa fa-mouse-pointer" aria-hidden="true"></i>
           </button>
-          <button type="button" onclick="setTool(moveMultiTool)" title="Move Multi Tool" class="toolButton">
+          <button type="button" onclick="setTool(drawMoveMultiTool)" title="Move Multi Tool" class="toolButton">
             <i class="fa fa-arrows" aria-hidden="true"></i>
           </button>
         </div>
